Move modal transition to modal component
